feat: Update clone_wp_site.py to clone src app config
Applications store their config in a 'json' field (which is submitted as JSON in the REST API).  In opalstack-python, this field is represented by a dictionary.
This change stores the source app config and applies it to the created app, which preserves such settings as "gzip", "php_version", and "auto_site_url".